sudoku/requirements.txt
typedb-client